pyproject.toml
[tool.poetry]
name = "telegrinder-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
telegrinder = "^0.1.dev9"
tortoise-orm = "^0.19.2"
aerich = "^0.7.1"
asyncpg = "^0.26.0"
envparse = "^0.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"
mypy = "^0.982"
autopep8 = "^1.7.0"


[tool.aerich]
tortoise_orm = "src.config.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 99

[tool.mypy]
ignore_missing_imports = true
follow_imports_for_stubs = false
ignore_missing_imports_per_module = true
